% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateS.R
\name{generateS}
\alias{generateS}
\title{generateS}
\usage{
generateS(seed = 15213, S = 3, Types = 2, k = 3,
  meanSamplesPerK = c(40, 40, 30), nModule = 30, meanGenesPerModule = 30,
  Gmean = 9, Gsd = 2, sigma1 = 1, sigma2 = 1, sigma3 = 1, G0 = 5000,
  nconfounder = 4, nrModule = 20, rMeanSubtypes = 3, diffmu = 1,
  fold = rep(1, S), rho = 0.5, df.prior = 100, groupProb = 1)
}
\arguments{
\item{seed}{random seed}

\item{S}{number of studies}

\item{Types}{number of omics types. I.e. Gene expression, DNA methylation}

\item{k}{number of clusters}

\item{meanSamplesPerK}{mean samples per cluster}

\item{nModule}{number of modules. A module is a group of genes.}

\item{meanGenesPerModule}{number of genes per module}

\item{Gmean}{gene expression template follows N(Gmean,Gsd^2)}

\item{Gsd}{gene expression template follows N(Gmean,Gsd^2)}

\item{sigma1}{noise 1}

\item{sigma2}{noise 2}

\item{sigma3}{noise 3}

\item{G0}{number of noise genes}

\item{nconfounder}{number of confounders}

\item{nrModule}{number of modules for confounding variables}

\item{rMeanSubtypes}{number of subtypes defined by confounding variables}

\item{diffmu}{effect size difference for subtype predictive genes}

\item{fold}{how to vary subtype predictive gene signal. 1: original. 0: no signal.}

\item{rho}{para for inverse Wishart distribution.}

\item{df.prior}{para for inverse Wishart distribution.}

\item{groupProb}{subtype predictive genes have prior group information. By prob 1-groupProb, the information will be altered.}
}
\value{
alist
}
\description{
generate S for MISKmeans
}
\details{
generate S for MISKmeans
}
\examples{
Sdata =generateS(seed=15213,S=3,Types=2, k=3)
#
length(Sdata)
dim(Sdata[[1]]$d)
sum(Sdata[[1]]$subPredictGeneUnion)
}
\author{
Caleb
}
